tasks/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "lara-tasks"
version = "0.1.0"
description = "LARA pipeline tasks"
readme = "README.md"
dependencies = [
    "numpy",
    "tqdm",
    "pydantic",
    "tiktoken",
    "openai",
    "google-cloud-vision",
    "grpcio",
    "boto3",
    "types-boto3",
    "pillow",
    "geopy",
    "matplotlib",
    "opencv-python>=4.8",
    "protobuf",
    "scikit_image",
    "scikit_learn",
    "Shapely",
    "utm",
    "parmap",
    "scipy"
]

[project.optional-dependencies]
segmentation=[
    # torch is left out here because detectron requires it be present before pip
    # install runs - it should be installed manually as a precondition
    # (see https://github.com/facebookresearch/detectron2/issues/4472)#
    # "torch==2.0.1",
    "detectron2 @ git+https://github.com/facebookresearch/detectron2.git",
    "torchvision==0.15.2",
    "transformers",
    "timm"
]
point = [
    "torch==2.0.1",
    "torchvision==0.15.2",
    "scipy",
    "ultralytics"
]
development = [
    "pytest",
    "pyright",
    "moto"
]

[tool.setuptools.packages]
find = {}
